ref(search): Extract Ask Seer step formatting
Share progress-step normalization and copy formatting with alternate loading presentations without changing the existing progress UI.

Refs BROWSE-617
